+++ Tests/Models/UniqueEntityTests.cs
@@ -1,0 +1,25 @@
+﻿//using Microsoft.VisualStudio.TestTools.UnitTesting;
+//using Models.Common;
+//using Models.Data;
+//using Util.Random;
+
+//namespace Tests.Models {
+
+//    [TestClass] public class UniqueEntityTests : AbstractClassTests<UniqueEntity<OrderData>, Entity<MeasureData>> {
+
+//        private class testClass : UniqueEntity<MeasureData> {
+
+//            public testClass(MeasureData d = null) : base(d) { }
+
+//        }
+
+//        [TestInitialize] public override void TestInitialize() {
+//            base.TestInitialize();
+//            obj = new testClass(GetRandom.Object<MeasureData>());
+//        }
+
+//        [TestMethod] public void IdTest() => isReadOnlyProperty(obj.Id);
+
+//    }
+
+//}
