--- Tests/Repositories/Orders/OrdersRepositoryTests.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Repositories;
-
-namespace Tests.Repositories.Orders {
-
-    [TestClass] public class OrdersRepositoryTests : SealedTests<OrderRepository, object> { }
-
-}
-
